LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/moving-fontface-between-document-throws.tentative.html

<!DOCTYPE html>
<title>Tests that moving a FontFace between documents throws an exception</title>
<link rel="help" href="https://drafts.csswg.org/css-font-loading/">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/13251">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
    <div style="font: 32px 'Ahem';">Hello</div>
    <iframe id="target" srcdoc="
    <!DOCTYPE html>
    <html>
    <head>
    </head>
    <body>
    <div style='font: 32px Ahem;'>Hello</div>
    </body>
    </html>"></iframe>
<script>
test(function(t) {
    const fontFace = new FontFace("Ahem", "url(/fonts/Ahem.ttf)");
    document.fonts.add(fontFace);
    const iframeWindow = document.getElementById("target").contentWindow;
    assert_throws_dom("WrongDocumentError", iframeWindow.DOMException, function () {
        iframeWindow.document.fonts.add(fontFace);
    });
}, "Adding a FontFace to the FontFaceSet in a different document throws");
</script>
</body>
